80% of refactoring
