+++ app/src/main/java/com/example/androidlab/PrimaryOvalButtonComponent.kt
@@ -1,0 +1,28 @@
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun PrimaryOvalButton(text: String, modifier: Modifier = Modifier) {
+    TextButton(
+        onClick = { },
+        modifier = modifier,
+        shape = RoundedCornerShape(20.dp),
+        colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFFFAC00))
+    ) {
+        Text(
+            text = text,
+            color = Color.Black,
+            fontSize = 20.sp,
+            fontWeight = FontWeight.W700,
+            letterSpacing = 0.6.sp
+        )
+    }
+}
