+++ app/src/main/java/com/example/androidlab/ui/theme/Type.kt
@@ -1,7 +1,10 @@
 package com.example.androidlab.ui.theme
 
+
 import androidx.compose.material3.Typography
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
